First version
Summary of the script's functionality and features:
Node Verification at Start:

Before starting any other process, the script checks the availability of the source node and the destination node. If both are available, it proceeds with the migration process. These modifications ensure that the script does not attempt to execute migration operations if any of the nodes are not accessible on the network, preventing potential errors and unexpected behavior.

*Note: Before running the script, make sure you have authorized access to the hosts using the command: SSH AUTHORIZATION

ssh-keygen -t rsa -b 4096 -C "root@local" <-- If you already have an id, this command might not be necessary.
ssh-copy-id root@<ip_host_1>
ssh-copy-id root@<ip_host_2>

This script automates the migration of a virtual machine (VM) between two nodes in a Proxmox environment, executed from a third Linux machine that accesses the nodes via SSH. The main features and functionality are as follows:

Remote Execution:

The script is run from an external control machine and uses SSH to execute all commands on the source and destination nodes, ensuring that the migration is done remotely and automatically.
Input Parameters:

Requires the VM ID as a mandatory argument.
Accepts two additional options: --force-shutdown to shut down the VM if it is running, and --snapshot to handle snapshot references during the migration process.
Snapshot and VM State Verification:

It checks if the VM has snapshots on the source node, and if so, it stops the process until they are removed.
It verifies the VM’s status and, if it’s running and --force-shutdown is specified, it gracefully shuts it down before proceeding.
Verification and Cleanup on the Destination Node:

Before migration, it checks if the VM already exists on the destination node. If associated disks exist, it removes them from the destination node only if they are still configured.
Copying VM Configuration and Disks:

Transfers the VM’s configuration file from the source node to the destination node using scp over SSH.
Copies the VM’s disks from the source node to the destination node using the pvesm export/import command through an SSH connection, displaying detailed progress for each disk.

Snapshot Handling:

If --snapshot is specified, it removes snapshot references in the VM configuration file on the destination node to ensure a clean configuration.
Completion:

Once the migration is complete, the script displays a success message, indicating that the process has been successfully completed.
The script is designed to run robustly and securely from an external control machine, ensuring that all necessary commands for VM migration are executed on the corresponding nodes via SSH. This facilitates the automation of migration in distributed environments.